--- a/src/main/scala/beam/agentsim/agents/ridehail/VehicleCentricMatchingForRideHail.scala
+++ b/src/main/scala/beam/agentsim/agents/ridehail/VehicleCentricMatchingForRideHail.scala
@@ -150,15 +150,15 @@ class VehicleCentricMatchingForRideHail(
   }
 
   private def greedyAssignment(trips: List[AssignmentKey]): List[AssignmentKey] = {
-    val Rok = mutable.ListBuffer.empty[CustomerRequest]
-    val Vok = mutable.ListBuffer.empty[VehicleAndSchedule]
+    val Rok = mutable.HashSet.empty[CustomerRequest]
+    val Vok = mutable.HashSet.empty[VehicleAndSchedule]
     val greedyAssignmentList = mutable.ListBuffer.empty[AssignmentKey]
     var tripsSorted = trips.sortBy(-_._1.requests.size)
     while (tripsSorted.nonEmpty) {
       tripsSorted.filter(_._1.requests.size == tripsSorted.head._1.requests.size).sortBy(_._3).foreach {
         case (trip, vehicle, cost) if (!Vok.contains(vehicle) && !trip.requests.exists(r => Rok.contains(r))) =>
-          Rok.appendAll(trip.requests)
-          Vok.append(vehicle)
+          trip.requests.foreach(Rok.add)
+          Vok.add(vehicle)
           greedyAssignmentList.append((trip, vehicle, cost))
         case _ =>
       }