--- a/src/main/scala/beam/agentsim/agents/ridehail/allocation/PoolingAlonsoMora.scala
+++ b/src/main/scala/beam/agentsim/agents/ridehail/allocation/PoolingAlonsoMora.scala
@@ -148,12 +148,14 @@ class PoolingAlonsoMora(val rideHailManager: RideHailManager)
               rideHailManager.beamServices,
               vehState.totalRemainingRange - rideHailManager.beamScenario.beamConfig.beam.agentsim.agents.rideHail.rangeBufferForDispatchInMeters
             )
-            rideHailManager.log.debug(
-              "%%%%% Vehicle {} is available with this schedule: \n {}",
-              vehAndSched.vehicle.id,
-              vehAndSched.schedule.map(_.toString).mkString("\n")
-            )
-            vehAndSched
+            if (rideHailManager.log.isDebugEnabled) {
+              rideHailManager.log.debug(
+                "%%%%% Vehicle {} is available with this schedule: \n {}",
+                vehAndSched.vehicle.id,
+                vehAndSched.schedule.map(_.toString).mkString("\n")
+              )
+            }
+              vehAndSched
           }.toList,
           pooledAllocationReqs.map(
             rhr =>
@@ -172,8 +174,10 @@ class PoolingAlonsoMora(val rideHailManager: RideHailManager)
       poolCustomerReqs.foreach { d =>
         spatialPoolCustomerReqs.put(d.pickup.activity.getCoord.getX, d.pickup.activity.getCoord.getY, d)
       }
-      rideHailManager.log
-        .debug("%%%%% Requests: {}", spatialPoolCustomerReqs.values().asScala.map(_.toString).mkString("\n"))
+      if (rideHailManager.log.isDebugEnabled) {
+        rideHailManager.log
+          .debug("%%%%% Requests: {}", spatialPoolCustomerReqs.values().asScala.map(_.toString).mkString("\n"))
+      }
       val alg =
         new VehicleCentricMatchingForRideHail(
           spatialPoolCustomerReqs,
@@ -310,10 +314,12 @@ class PoolingAlonsoMora(val rideHailManager: RideHailManager)
       e = System.currentTimeMillis()
       logger.debug(s"Served soloCustomer ${soloCustomer.size} in ${e - s} ms")
     }
-    rideHailManager.log.debug(
-      "AllocResponses: {}",
-      allocResponses.groupBy(_.getClass).map(x => s"${x._1.getSimpleName} -- ${x._2.size}").mkString("\t")
-    )
+    if (rideHailManager.log.isDebugEnabled) {
+      rideHailManager.log.debug(
+        "AllocResponses: {}",
+        allocResponses.groupBy(_.getClass).map(x => s"${x._1.getSimpleName} -- ${x._2.size}").mkString("\t")
+      )
+    }
     VehicleAllocations(allocResponses)
   }
 }