--- a/src/main/scala/beam/agentsim/agents/ridehail/allocation/PoolingAlonsoMora.scala
+++ b/src/main/scala/beam/agentsim/agents/ridehail/allocation/PoolingAlonsoMora.scala
@@ -155,7 +155,7 @@ class PoolingAlonsoMora(val rideHailManager: RideHailManager)
                 vehAndSched.schedule.map(_.toString).mkString("\n")
               )
             }
-              vehAndSched
+            vehAndSched
           }.toList,
           pooledAllocationReqs.map(
             rhr =>