--- a/src/main/scala/beam/agentsim/agents/ridehail/allocation/RandomRepositioning.scala
+++ b/src/main/scala/beam/agentsim/agents/ridehail/allocation/RandomRepositioning.scala
@@ -3,6 +3,7 @@ package beam.agentsim.agents.ridehail.allocation
 import java.io.{File, FileWriter}
 
 import beam.agentsim.agents.ridehail.RideHailManager
+import beam.agentsim.agents.ridehail.RideHailVehicleManager.RideHailAgentLocation
 import beam.agentsim.agents.ridehail.repositioningmanager.DemandFollowingRepositioningManager
 import beam.analysis.plots.GraphsStatsAgentSimEventsListener
 import beam.router.BeamRouter.Location
@@ -192,7 +193,7 @@ class RandomRepositioning(val rideHailManager: RideHailManager)
   // Map from tick to the pair of vehicleId (who to reposition) and location (where).
   var tickToLocation: Map[Int, Vector[(Id[Vehicle], Location)]] = Map.empty
 
-  override def repositionVehicles(tick: Int): Vector[(Id[Vehicle], Location)] = {
+  override def repositionVehicles(idleVehicles: scala.collection.Map[Id[Vehicle], RideHailAgentLocation], tick: Int): Vector[(Id[Vehicle], Location)] = {
 
 //    VehicleShouldRefuel(Id,Option[RefuelingLocation]) // None -> signal the human to refuel
 
@@ -204,7 +205,7 @@ class RandomRepositioning(val rideHailManager: RideHailManager)
 
     algorithm match {
       case 8 =>
-        algo8.repositionVehicles(tick)
+        algo8.repositionVehicles(Map.empty, tick)
       // TODO: destinations of idle vehicles selected for repositioning should be uniformally distributed in activity space
 
       // This should perform the same as DEFAULT_MANAGER!!!