--- a/src/main/scala/beam/agentsim/agents/ridehail/allocation/RandomRepositioning.scala
+++ b/src/main/scala/beam/agentsim/agents/ridehail/allocation/RandomRepositioning.scala
@@ -193,7 +193,10 @@ class RandomRepositioning(val rideHailManager: RideHailManager)
   // Map from tick to the pair of vehicleId (who to reposition) and location (where).
   var tickToLocation: Map[Int, Vector[(Id[Vehicle], Location)]] = Map.empty
 
-  override def repositionVehicles(idleVehicles: scala.collection.Map[Id[Vehicle], RideHailAgentLocation], tick: Int): Vector[(Id[Vehicle], Location)] = {
+  override def repositionVehicles(
+    idleVehicles: scala.collection.Map[Id[Vehicle], RideHailAgentLocation],
+    tick: Int
+  ): Vector[(Id[Vehicle], Location)] = {
 
 //    VehicleShouldRefuel(Id,Option[RefuelingLocation]) // None -> signal the human to refuel