--- a/src/main/scala/beam/agentsim/agents/vehicles/BeamVehicle.scala
+++ b/src/main/scala/beam/agentsim/agents/vehicles/BeamVehicle.scala
@@ -332,7 +332,7 @@ class BeamVehicle(
     }
   }
 
-  override def toString = s"$id ($beamVehicleType.id)"
+  override def toString = s"$id (${beamVehicleType.id})"
 }
 
 object BeamVehicle {