## Main changes in PR:
- Replace data structures to appropriate one to reduce runtime
- Explicitly put check for the log level because Actor Logging need to calculate args before calling logging method
- Intensive operation to get idle vehicles is cached and passed instead of re-computing it

## Run params:
- Config: `production/sfbay/smart/smart-b-hightech-2040.conf`, for 5 iterations
- Baseline: branch [art/#2223-prod-dev-updated](https://github.com/LBNL-UCB-STI/beam/tree/art/%232223-prod-dev-updated), commit 674b838aefac6c9ed16de6bfa3fd17999b8235e2, result: https://s3.us-east-2.amazonaws.com/beam-outputs/index.html#output/sfbay/sfbay-smart-b-ht-2040__2019-10-25_19-37-05
- Optimized: branch [art/#2223-prod-dev-updated](https://github.com/LBNL-UCB-STI/beam/tree/art/%232223-prod-dev-updated), commit 9c0cb1fe62b2a1783aded2fdeddba493c820971a, result: https://s3.us-east-2.amazonaws.com/beam-outputs/index.html#output/sfbay/sfbay-smart-b-ht-2040__2019-10-27_09-01-07/

## Results
Run | Iter 0, s | Iter 1, s | Iter 2, s | Iter 3, s | Iter 4, s | Iter 5, s | Sum, s
-- | -- | -- | -- | -- | -- | -- | --
Baseline | 4212 | 3145 | 3203 | 3102 | 3253 | 3133 | 20048
Optimized | 2775 | 1927 | 1886 | 1895 | 1901 | 2005 | 12389
Baseline /   Optimized | 1.52 | 1.63 | 1.70 | 1.64 | 1.71 | 1.56 | 1.62

## Profiling tool results
### Baseline overall
![image](https://user-images.githubusercontent.com/5107562/67637456-80a99800-f90d-11e9-9cf8-1ad83d23fb27.png)
### Baseline 1-th iteration
![image](https://user-images.githubusercontent.com/5107562/67637520-f7469580-f90d-11e9-9e5c-b1fa2114a610.png)

### Optimized overall
![image](https://user-images.githubusercontent.com/5107562/67637558-442a6c00-f90e-11e9-88ec-c48259feccaa.png)
### Optimized 1-th iteration
![image](https://user-images.githubusercontent.com/5107562/67637600-9c616e00-f90e-11e9-99d1-f68cf63f73e0.png)





<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/lbnl-ucb-sti/beam/2242)
<!-- Reviewable:end -->